requirements.txt
absl-py==2.1.0
altair==5.4.1
apturl==0.5.2
astunparse==1.6.3
attrs==24.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
fasteners==0.14.1
flatbuffers==24.3.25
future==0.18.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
namex==0.0.8
narwhals==1.12.1
netifaces==0.11.0
numpy==2.0.2
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.28.3
ptyprocess==0.7.0
pyarrow==18.0.0
pycairo==1.20.1
pycups==2.0.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rich==13.9.3
rpds-py==0.20.1
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
systemd-python==234
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
watchdog==5.0.3
Werkzeug==3.1.0
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
